refactor : 로그인 세션 정보 조건을 추가해서 기능 수정
-showQuestionForm(): 로그인 세션 정보가 있을 경우에만 질문 폼 띄워지도록 수정
-showEditForm(): 질문 수정폼 매핑
-create(): 사용자가 로그인했을 경우에만 질문 작성 가능하도록 숮어
-edit(): 작성자와 로그인한 사용자가 일치할 경우에만 질문내용 수정 가능하도록 수정
-remove(): 작성자와 로그인한 사용자가 일치할 경우에만 질문 삭제 가능하도록 수정

+++ src/main/java/codesquad/controller/QuestionController.java
@@ -1,12 +1,16 @@
 package codesquad.controller;
 
 import codesquad.domain.Question;
+import codesquad.domain.User;
+import codesquad.domain.WriteQuestionDto;
 import codesquad.exception.NoSuchQuestionException;
+import codesquad.exception.NoSuchUserException;
 import codesquad.repository.QuestionRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.time.LocalDateTime;
 
 @Controller
@@ -18,15 +22,13 @@ public class QuestionController {
         this.questionRepository = questionRepository;
     }
 
-    @PostMapping("/questions")
-    public String create(@ModelAttribute Question question) {
-        question.createWrittenTime(getWrittenTime());
-        questionRepository.save(question);
-        return "redirect:/questions";
-    }
-
-    private LocalDateTime getWrittenTime() {
-        return LocalDateTime.now();
+    @GetMapping("/questions/form")
+    public String showQuestionForm(HttpSession session) {
+        Object value = session.getAttribute("sessionedUser");
+        if (value != null) {
+            return "qna/form";
+        }
+        return "qna/no_session_form";
     }
 
     @GetMapping("/questions")
@@ -43,10 +45,64 @@ public class QuestionController {
         return "qna/show";
     }
 
+    @GetMapping("/questions/{id}/editor")
+    public String showEditForm() {
+        return "qna/edit_form";
+    }
+
+    @PostMapping("/questions")
+    public String create(@ModelAttribute WriteQuestionDto writeQuestionDto, HttpSession session) {
+        Object value = session.getAttribute("sessionedUser");
+        if (value != null) {
+            Question question = new Question();
+            User sessionedUser = (User) value;
+            question.write(sessionedUser, writeQuestionDto, getWrittenTime());
+            questionRepository.save(question);
+            return "redirect:/questions";
+        }
+        return "qna/no_session_form";
+    }
+
+    private LocalDateTime getWrittenTime() {
+        return LocalDateTime.now();
+    }
+
+    @PutMapping("/questions/{id}")
+    public String edit(@PathVariable Long id, WriteQuestionDto writeQuestionDto, HttpSession session) {
+        Object value = session.getAttribute("sessionedUser");
+        // 작성 부분으로 이동시키고 작성해서 다시 돌아오게 만들어야함
+        if (value != null) {
+            try {
+                User sessionedUser = (User) value;
+                Question question = questionRepository.findById(id)
+                                                      .filter(s -> s.getId()
+                                                                    .equals(sessionedUser.getId()))
+                                                      .orElseThrow(NoSuchUserException::new
+                                                      );
+
+                question.write(sessionedUser, writeQuestionDto, getWrittenTime());
+                questionRepository.save(question);
+            } catch (NoSuchUserException exception) {
+                return "qna/no_access_form";
+            }
+        }
+        return "redirect:/questions";
+    }
+
     @DeleteMapping("/questions/{id}")
-    public String remove(@PathVariable Long id) {
-        questionRepository.delete(questionRepository.findById(id)
-                                                    .orElseThrow(NoSuchQuestionException::new));
+    public String remove(@PathVariable Long id, HttpSession session) {
+        Object value = session.getAttribute("sessionedUser");
+        if (value != null) {
+            try {
+                User sessionedUser = (User) value;
+                questionRepository.delete(questionRepository.findById(id)
+                                                            .filter(s -> s.getId()
+                                                                          .equals(sessionedUser.getId()))
+                                                            .orElseThrow(NoSuchQuestionException::new));
+            } catch (NoSuchQuestionException exception) {
+                return "redirect:/users/login";
+            }
+        }
         return "redirect:/questions";
     }
 }
